Reorganization and CameraController improvements
